Further optimise day09
To further optimise this solution, checking whether the tail position was already in `tail_positions` before adding it to the list was removed, meaning the full list of positions didn't need to be traversed each time the tracked tail moved.

Instead, all tail positions of the tracked tail, including duplicates, are added as a tuple to the list. This allows to convert the list to a set at the end of the function, removing duplicates and leaving only the unique positions of the tracked tail.

This optimisation reduced execution time by ~86% (0.36s -> 0.05s).

All optimisations thus far have resulted in a ~94% decrease in execution time (0.9s -> 0.05s).